fix: consolidate open issues/PRs for OpenVX 1.3 sample impl
Fixes included:
- #20: vxCreateThresholdForImage returns error object on bad params
- #48: vxRead/WriteScalarValue respects data_addr for WithSize scalars
- #52: release virtual objects scoped to graph on vxReleaseGraph
- #21/#22: vxMapTensorPatch strides based on full tensor dimensions
- #23/#24: PoolingKernelImpl signed/unsigned cast fix
- #6: XML export threshold true/false_value typed correctly (S32)
- #50/#29: remove -mfpu=neon for aarch64 builds
- #46: add /bigobj to Windows debug flags
- #40: guard FIND_NUM_PROCESSORS during cross-compilation
- #45: add --userdataobj flag to Build.py
- #54: rename reserved identifier header guards
- #59: move Laplacian Pyramid/Reconstruct compute from Initializer to Kernel

Refs: #6 #20 #21 #22 #23 #24 #29 #30 #34 #40 #45 #46 #48 #50 #52 #53 #54 #59

+++ kernels/c_model/c_khr_nn.c
@@ -508,7 +508,7 @@ void PoolingKernelImpl(
         if (!max_pooling)
         {
             //result = conversion_24_8(result / (int16_t)(size_x * size_y));
-            result = CLAMP(result / (size_x * size_y), getMinValue(fmt), getMaxValue(fmt));
+            result = CLAMP(result / (int32_t)(size_x * size_y), getMinValue(fmt), getMaxValue(fmt));
         }
 
         const size_t output_byte_offset =
